+++ fiddles/bash/fiddle-0093-InitAndroidProject/run.sh
@@ -1,16 +1,14 @@
 #!/usr/bin/env bash
-
-source ../../../scripts/bin/_utils.sh;
-
+source scripts/bin/_utils.sh;
 
 clear;
-echo "$0" | sed 's/\.\///g' | awk '{print toupper($0)}'
-echo "Bash version ${BASH_VERSION}..."
+echo "$0" | sed 's/\.\///g' | awk '{print toupper($0)}';
+echo "Bash version ${BASH_VERSION}...";
 
 
 function catch() {
     case $1 in
-        0)  endLog "environment configured";
+        0)  endLog "$0";
             ;;
         *)  echo "fubar! Something went wrong."
             ;;
@@ -19,9 +17,9 @@ function catch() {
 }
 # try
 (
-    _path="fiddle";
-    if [ "$#" -eq 1 ]; then _path=$1; fi
-    # Do stuff
+     _path="fiddle-0001-Console";
+    cd scripts;
+    ./fiddle-android.sh ${_path};
 
 )
 catch $?;