+++ fiddles/bash/fiddle-0093-InitAndroidProject/scripts/bin/_utils.sh
@@ -23,6 +23,11 @@ function split() {
   echo ${str//$delimiter/ };
 }
 
+function capitalize() {
+    A=$1;
+    for i in $A; do B=`echo -n "${i:0:1}" | tr "[:lower:]" "[:upper:]"`; echo -n "${B}${i:1} "; done
+}
+
 function parseName() {
   pieces=($(split $1 "-";));
   if [[ ${#pieces[@]} -eq 3 ]]