+++ fiddles/bash/fiddle-0093-InitAndroidProject/test.sh
@@ -7,8 +7,18 @@ source scripts/bin/android/_install.sh;
 source scripts/bin/android/_start.sh;
 source scripts/bin/android/_fork.sh;
 
-function testIsJavaInstalled() {
-  actual=$(isJavaInstalled;);
+function testUtilCapitalize() {
+  actual=$(capitalize "template";);
+  expected="Template";
+  echo ${actual};
+  if [[ ${actual} -ne ${expected} ]]
+  then
+    exit 5;
+  fi
+}
+
+function testIsAndroidInstalled() {
+  actual=$(isAndroidInstalled;);
   expected=0;
   if [[ ${actual} -ne ${expected} ]]
   then
@@ -16,78 +26,39 @@ function testIsJavaInstalled() {
   fi
 }
 
-function testIsBrewInstalled() {
-  actual=$(isBrewInstalled;);
-  expected=0;
-  if [[ ${actual} -ne ${expected} ]]
-  then
-    exit 4;
-  fi
-}
-
-function testIsSdkInstalled() {
-  actual=$(isSdkInstalled;);
-  expected=0;
-  if [[ ${actual} -ne ${expected} ]]
-  then
-    exit 3;
-  fi
-}
-
-function testIsGradleInstalled() {
-  actual=$(isGradleInstalled;);
-  expected=0;
-  if [[ ${actual} -ne ${expected} ]]
-  then
-    exit 2;
-  fi
-}
-
-function testIsGitInstalled() {
-  actual=$(isGitInstalled;);
-  expected=0;
-  if [[ ${actual} -ne ${expected} ]]
-  then
-    exit 5;
-  fi
-}
-
-function testJavaCreate() {
+function testAndroidCreate() {
   bornOnDate=$(date +"%m-%d-%y";)
   name="fiddle-0000-Template";
   suffix=$(parseName ${name};);
   projectName=$(toLowerCase ${suffix};);
-  cd "fiddles/java";
-  javaCreate ${name} ${bornOnDate} ${projectName} ${__SEEDER__} || exit 7;
-  cd "../..";
+  cd "fiddles/android";
+  androidCreate ${name} ${bornOnDate} ${projectName} || exit 2;
 }
 
-function testJavaStart() {
+function testAndroidStart() {
   name="fiddle-0000-Template";
-  cd "fiddles/java/${name}";
-  javaStart || exit 8;
-  cd "../../../..";
+  cd "fiddles/android/${name}";
+  androidStart || exit 3;
 }
 
-function testJavaFork() {
+function testAndroidFork() {
   origFiddle="fiddle-0000-Template";
-  newFiddle="fiddle-0001-TemplateFork";
+  newFiddle="fiddle-0001-Fork";
   cd scripts;
-  if [[ -d "../fiddles/java/${newFiddle}" ]]; then rm -rf "../fiddles/java/${newFiddle}" || exit 9; fi
-  cp -rf "../fiddles/java/${origFiddle}" "../fiddles/java/${newFiddle}" || exit 9;
-  cd "../fiddles/java";
-  javaFork ${origFiddle} ${newFiddle} || exit 9;
-  cd "../..";
+  if [[ -d "../fiddles/android/${newFiddle}" ]]; then rm -rf "../fiddles/android/${newFiddle}" || exit 4; fi
+  cp -rf "../fiddles/android/${origFiddle}" "../fiddles/android/${newFiddle}" || exit 4;
+  cd "../fiddles/android";
+  androidFork ${origFiddle} ${newFiddle} || exit 4;
 }
 
 function tearDown() {
-  if [[ -d "./fiddles/java/fiddle-0000-Template" ]]
+  if [[ -d "./fiddles/android/fiddle-0000-Template" ]]
   then
-    rm -rf "./fiddles/java/fiddle-0000-Template";
+    rm -rf "./fiddles/android/fiddle-0000-Template";
   fi
-  if [[ -d "./fiddles/java/fiddle-0001-TemplateFork" ]]
+  if [[ -d "./fiddles/android/fiddle-0001-Fork" ]]
   then
-    rm -rf "./fiddles/java/fiddle-0001-TemplateFork";
+    rm -rf "./fiddles/android/fiddle-0001-Fork";
   fi
 }
 
@@ -95,23 +66,15 @@ function catch() {
     case $1 in
         0)  endLog $(echo "$0" | sed 's/\.\///g');
             ;;
-        1)  endLog "test 1 failed - testIsJavaInstalled";
+        1)  endLog "test 1 failed - testIsAndroidInstalled";
             ;;
-        2)  endLog "test 2 failed - testIsGradleInstalled";
+        2)  endLog "test 2 failed - testAndroidCreate";
             ;;
-        3)  endLog "test 3 failed - testIsSdkInstalled";
+        3)  endLog "test 3 failed - testAndroidStart";
             ;;
-        4)  endLog "test 4 failed - testIsBrewInstalled";
+        4)  endLog "test 4 failed - testAndroidFork";
             ;;
-        5)  endLog "test 5 failed - testIsGitInstalled";
-            ;;
-        6)  endLog "test 6 failed - testGitCloneSeeder";
-            ;;
-        7)  endLog "test 7 failed - testJavaCreate";
-            ;;
-        8)  endLog "test 8 failed - testJavaStart";
-            ;;
-        9)  endLog "test 9 failed - testJavaFork";
+        5)  endLog "test 5 failed - testUtilCapitalize";
             ;;
         *)  echo "fubar! Something went wrong."
             ;;
@@ -120,25 +83,23 @@ function catch() {
 }
 # try
 (
+    startingPoint=$(pwd);
     startLog $(echo "$0" | sed 's/\.\///g');
-    groupLog "testIsJavaInstalled";
-    testIsJavaInstalled || exit $?;
-    groupLog "testIsGradleInstalled";
-    testIsGradleInstalled || exit $?;
-    groupLog "testIsSdkInstalled";
-    testIsSdkInstalled || exit $?;
-    groupLog "testIsBrewInstalled";
-    testIsBrewInstalled || exit $?;
-    groupLog "testIsGitInstalled";
-    testIsGitInstalled || exit $?;
-    groupLog "testJavaCreate";
-    testJavaCreate || exit $?;
-    groupLog "testJavaStart";
-    testJavaStart || exit $?;
-    groupLog "testJavaFork";
-    testJavaFork || exit $?;
+    groupLog "testUtilCapitalize";
+    testUtilCapitalize || exit $?;
+    groupLog "testIsAndroidInstalled";
+    testIsAndroidInstalled || exit $?;
+    groupLog "testAndroidCreate";
+    testAndroidCreate || exit $?;
+    cd "${startingPoint}";
+    groupLog "testAndroidStart";
+    testAndroidStart || exit $?;
+    cd "${startingPoint}";
+    groupLog "testAndroidFork";
+    testAndroidFork || exit $?;
+    cd "${startingPoint}";
     groupLog "tearDown";
-    tearDown;
+    #tearDown;
 )
 rc=$?; catch ${rc};
 # finally